refactor: create SidebarContext
