+++ src/library-authoring/common/context/SidebarContext.tsx
@@ -1,0 +1,164 @@
+import {
+  createContext,
+  useCallback,
+  useContext,
+  useMemo,
+  useState,
+} from 'react';
+
+export enum SidebarBodyComponentId {
+  AddContent = 'add-content',
+  Info = 'info',
+  ComponentInfo = 'component-info',
+  CollectionInfo = 'collection-info',
+}
+
+export const COLLECTION_INFO_TABS = {
+  Manage: 'manage',
+  Details: 'details',
+} as const;
+export type CollectionInfoTab = typeof COLLECTION_INFO_TABS[keyof typeof COLLECTION_INFO_TABS];
+export const isCollectionInfoTab = (tab: string): tab is CollectionInfoTab => (
+  Object.values<string>(COLLECTION_INFO_TABS).includes(tab)
+);
+
+export const COMPONENT_INFO_TABS = {
+  Preview: 'preview',
+  Manage: 'manage',
+  Details: 'details',
+} as const;
+export type ComponentInfoTab = typeof COMPONENT_INFO_TABS[keyof typeof COMPONENT_INFO_TABS];
+export const isComponentInfoTab = (tab: string): tab is ComponentInfoTab => (
+  Object.values<string>(COMPONENT_INFO_TABS).includes(tab)
+);
+
+export interface SidebarComponentInfo {
+  type: SidebarBodyComponentId;
+  id: string;
+  /** Additional action on Sidebar display */
+  additionalAction?: SidebarAdditionalActions;
+  /** Current tab in the sidebar */
+  currentTab?: CollectionInfoTab | ComponentInfoTab;
+}
+
+export enum SidebarAdditionalActions {
+  JumpToAddCollections = 'jump-to-add-collections',
+}
+
+export type SidebarContextData = {
+  closeLibrarySidebar: () => void;
+  openAddContentSidebar: () => void;
+  openInfoSidebar: () => void;
+  openCollectionInfoSidebar: (collectionId: string, additionalAction?: SidebarAdditionalActions) => void;
+  openComponentInfoSidebar: (usageKey: string, additionalAction?: SidebarAdditionalActions) => void;
+  sidebarComponentInfo?: SidebarComponentInfo;
+  resetSidebarAdditionalActions: () => void;
+  setSidebarCurrentTab: (tab: CollectionInfoTab | ComponentInfoTab) => void;
+};
+
+/**
+ * Sidebar Context.
+ *
+ * Get this using `useSidebarContext()`
+ *
+ */
+const SidebarContext = createContext<SidebarContextData | undefined>(undefined);
+
+type SidebarProviderProps = {
+  children?: React.ReactNode;
+  /** Only used for testing */
+  initialSidebarComponentInfo?: SidebarComponentInfo;
+};
+
+/**
+ * React component to provide `LibraryContext`
+ */
+export const SidebarProvider = ({
+  children,
+  initialSidebarComponentInfo,
+}: SidebarProviderProps) => {
+  const [sidebarComponentInfo, setSidebarComponentInfo] = useState<SidebarComponentInfo | undefined>(
+    initialSidebarComponentInfo,
+  );
+
+  /** Helper function to consume addtional action once performed.
+    Required to redo the action.
+  */
+  const resetSidebarAdditionalActions = useCallback(() => {
+    setSidebarComponentInfo((prev) => (prev && { ...prev, additionalAction: undefined }));
+  }, []);
+
+  const closeLibrarySidebar = useCallback(() => {
+    setSidebarComponentInfo(undefined);
+  }, []);
+  const openAddContentSidebar = useCallback(() => {
+    setSidebarComponentInfo({ id: '', type: SidebarBodyComponentId.AddContent });
+  }, []);
+  const openInfoSidebar = useCallback(() => {
+    setSidebarComponentInfo({ id: '', type: SidebarBodyComponentId.Info });
+  }, []);
+
+  const openComponentInfoSidebar = useCallback((usageKey: string, additionalAction?: SidebarAdditionalActions) => {
+    setSidebarComponentInfo((prev) => ({
+      ...prev,
+      id: usageKey,
+      type: SidebarBodyComponentId.ComponentInfo,
+      additionalAction,
+    }));
+  }, []);
+
+  const openCollectionInfoSidebar = useCallback((
+    newCollectionId: string,
+    additionalAction?: SidebarAdditionalActions,
+  ) => {
+    setSidebarComponentInfo((prev) => ({
+      ...prev,
+      id: newCollectionId,
+      type: SidebarBodyComponentId.CollectionInfo,
+      additionalAction,
+    }));
+  }, []);
+
+  const setSidebarCurrentTab = useCallback((tab: CollectionInfoTab | ComponentInfoTab) => {
+    setSidebarComponentInfo((prev) => (prev && { ...prev, currentTab: tab }));
+  }, []);
+
+  const context = useMemo<SidebarContextData>(() => {
+    const contextValue = {
+      closeLibrarySidebar,
+      openAddContentSidebar,
+      openInfoSidebar,
+      openComponentInfoSidebar,
+      sidebarComponentInfo,
+      openCollectionInfoSidebar,
+      resetSidebarAdditionalActions,
+      setSidebarCurrentTab,
+    };
+
+    return contextValue;
+  }, [
+    closeLibrarySidebar,
+    openAddContentSidebar,
+    openInfoSidebar,
+    openComponentInfoSidebar,
+    sidebarComponentInfo,
+    openCollectionInfoSidebar,
+    resetSidebarAdditionalActions,
+    setSidebarCurrentTab,
+  ]);
+
+  return (
+    <SidebarContext.Provider value={context}>
+      {children}
+    </SidebarContext.Provider>
+  );
+};
+
+export function useSidebarContext(): SidebarContextData {
+  const ctx = useContext(SidebarContext);
+  if (ctx === undefined) {
+    /* istanbul ignore next */
+    throw new Error('useSidebarContext() was used in a component without a <SidebarProvider> ancestor.');
+  }
+  return ctx;
+}
