+++ src/library-authoring/component-info/ComponentInfo.tsx
@@ -11,16 +11,15 @@ import {
   CheckBoxOutlineBlank,
 } from '@openedx/paragon/icons';
 
+import { useComponentPickerContext } from '../common/context/ComponentPickerContext';
+import { useLibraryContext } from '../common/context/LibraryContext';
 import {
-  useComponentPickerContext,
-} from '../common/context/ComponentPickerContext';
-import {
-  SidebarAdditionalActions,
-  useLibraryContext,
+  type ComponentInfoTab,
   COMPONENT_INFO_TABS,
-  ComponentInfoTab,
+  SidebarAdditionalActions,
   isComponentInfoTab,
-} from '../common/context/LibraryContext';
+  useSidebarContext
+} from '../common/context/SidebarContext';
 import ComponentMenu from '../components';
 import { canEditComponent } from '../components/ComponentEditorModal';
 import ComponentDetails from './ComponentDetails';
@@ -34,9 +33,7 @@ import { ToastContext } from '../../generic/toast-context';
 const AddComponentWidget = () => {
   const intl = useIntl();
 
-  const {
-    sidebarComponentInfo,
-  } = useLibraryContext();
+  const { sidebarComponentInfo } = useSidebarContext();
 
   const {
     componentPickerMode,
@@ -103,13 +100,8 @@ const AddComponentWidget = () => {
 const ComponentInfo = () => {
   const intl = useIntl();
 
-  const {
-    sidebarComponentInfo,
-    readOnly,
-    openComponentEditor,
-    resetSidebarAdditionalActions,
-    setSidebarCurrentTab,
-  } = useLibraryContext();
+  const { readOnly, openComponentEditor } = useLibraryContext();
+  const { setSidebarCurrentTab, sidebarComponentInfo, resetSidebarAdditionalActions } = useSidebarContext();
 
   const jumpToCollections = sidebarComponentInfo?.additionalAction === SidebarAdditionalActions.JumpToAddCollections;
