+++ src/library-authoring/component-info/ManageCollections.test.tsx
@@ -13,6 +13,7 @@ import { mockContentSearchConfig } from '../../search-manager/data/api.mock';
 import { mockContentLibrary, mockLibraryBlockMetadata } from '../data/api.mocks';
 import ManageCollections from './ManageCollections';
 import { LibraryProvider } from '../common/context/LibraryContext';
+import { SidebarProvider } from '../common/context/SidebarContext';
 import { getLibraryBlockCollectionsUrl } from '../data/api';
 
 let axiosMock: MockAdapter;
@@ -24,7 +25,11 @@ mockContentSearchConfig.applyMock();
 
 const render = (ui: React.ReactElement) => baseRender(ui, {
   extraWrapper: ({ children }) => (
-    <LibraryProvider libraryId={mockContentLibrary.libraryId}>{children}</LibraryProvider>
+    <LibraryProvider libraryId={mockContentLibrary.libraryId}>
+      <SidebarProvider>
+        {children}
+      </SidebarProvider>
+    </LibraryProvider>
   ),
 });
