+++ src/library-authoring/components/ComponentCard.test.tsx
@@ -6,6 +6,7 @@ import {
   initializeMocks,
 } from '../../testUtils';
 import { LibraryProvider } from '../common/context/LibraryContext';
+import { SidebarProvider } from '../common/context/SidebarContext';
 import { getClipboardUrl } from '../../generic/data/api';
 import { ContentHit } from '../../search-manager';
 import ComponentCard from './ComponentCard';
@@ -46,7 +47,13 @@ const clipboardBroadcastChannelMock = {
 
 const libraryId = 'lib:org1:Demo_Course';
 const render = () => baseRender(<ComponentCard contentHit={contentHit} />, {
-  extraWrapper: ({ children }) => <LibraryProvider libraryId={libraryId}>{ children }</LibraryProvider>,
+  extraWrapper: ({ children }) => (
+    <LibraryProvider libraryId={libraryId}>
+      <SidebarProvider>
+        { children }
+      </SidebarProvider>
+    </LibraryProvider>
+  ),
 });
 
 describe('<ComponentCard />', () => {
