+++ src/library-authoring/legacy-integration/PreviewChangesEmbed.tsx
@@ -2,8 +2,6 @@ import { useIntl } from '@edx/frontend-platform/i18n';
 import { useParams, useSearchParams } from 'react-router-dom';
 import { Helmet } from 'react-helmet';
 
-import { LibraryProvider } from '../common/context/LibraryContext';
-import { getLibraryId } from '../../generic/key-utils';
 import CompareChangesWidget from '../component-comparison/CompareChangesWidget';
 import { useLibraryBlockMetadata } from '../data/apiHooks';
 import messages from '../component-comparison/messages';
@@ -28,16 +26,15 @@ const PreviewChangesEmbed = () => {
   }
   const [queryString] = useSearchParams();
   const oldVersion = parseInt(queryString.get('old') ?? '', 10) || 'published';
-  const libraryId = getLibraryId(usageKey);
   const { data: metadata } = useLibraryBlockMetadata(usageKey);
 
   return (
-    <LibraryProvider libraryId={libraryId}>
+    <>
       {/* It's not necessary since this will usually be in an <iframe>,
           but it's good practice to set a title for any top level page */}
       <Helmet><title>{intl.formatMessage(messages.iframeTitlePrefix)} | {metadata?.displayName ?? ''} | {process.env.SITE_NAME}</title></Helmet>
       <CompareChangesWidget usageKey={usageKey} oldVersion={oldVersion} newVersion="published" />
-    </LibraryProvider>
+    </>
   );
 };
